Build and run on Apple Silicon (macOS 26 / Xcode 26)
Drop the QuickTime.framework link, bump deployment target to 10.13,
clear hard-coded code-sign identities, fix a comparator return-type
mismatch in GCSerie/GCFrame, add NSPoint casts in GCFrame to
disambiguate -offset, and bypass the dead NSMovie path in
GCView -setURL:asUser: that crashed every file open.

Replace NSArchiver/NSUnarchiver with NSKeyedArchiver/NSKeyedUnarchiver
across all archiving sites, and switch struct encoding (NSPoint, NSRect)
to NSValue + encodeObject: so keyed encoding accepts it. Opt the app
delegate into secure restorable state, and break a responder-chain
recursion in GCDocument's forwarding category by refusing to forward
validRequestorForSendType:returnType: to mView.

Note: the archive format change is not backwards-compatible with
documents saved by previous GraphClick versions.

+++ Source/GCSerie.m
@@ -186,7 +186,7 @@
 	[point release];
 }
 
-int compare(id ptA, id ptB, void *info)
+NSInteger compare(id ptA, id ptB, void *info)
 {
 	float a = 0, b = 0;
 	switch (((int *)info)[0]) {
